+++ CHANGELOG.md
@@ -7,7 +7,14 @@
 ### Bugs fixed
 
 ### Enhancements
-- GLS-344 - Added restriction on document upload by file types and virus scanning from Wagtail admin
+
+## [2.11.0](https://github.com/uktrade/great-cms/releases/tag/2.11.0)
+
+[Full Changelog](https://github.com/uktrade/great-cms/compare/2.10.0...2.11.0)
+
+### Enhancements
+
+* GLS-344 - Added restriction on document upload by file types and virus scanning from Wagtail admin
 
 ## [2.10.0](https://github.com/uktrade/great-cms/releases/tag/2.10.0)